TODO @gent: --remove user from group --edit group (!= administrator, helpdesk, manager) --findByCn throws error if cn not found => how to catch error ?

+++ secureLDAP/src/main/java/co/ntier/secureLDAP/controller/HomeController.java
@@ -1,10 +1,11 @@
 package co.ntier.secureLDAP.controller;
 
 import java.io.IOException;
+import java.security.SecureRandom;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
+import java.util.Random;
 
 import javax.naming.InvalidNameException;
 import javax.naming.Name;
@@ -12,7 +13,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpRequest;
 import org.springframework.ldap.NamingException;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -34,14 +34,42 @@ public class HomeController {
 	private OdmGroupService groupService;
 
 	private String defaultView = "users";
-	private String errorMsg = "";
-	private String defaultPassword = "abc123456";
+	private String errorMsg;
+
+	// Source :
+	// http://www.java2s.com/Code/Java/Security/GeneratearandomStringsuitableforuseasatemporarypassword.htm
+	private static final Random RANDOM = new SecureRandom();
+	/** Length of password. @see #generateRandomPassword() */
+	public static final int PASSWORD_LENGTH = 8;
+
+	/**
+	 * Generate a random String suitable for use as a temporary password.
+	 * 
+	 * @return String suitable for use as a temporary password
+	 */
+	public static String generateRandomPassword() {
+		// Pick from some letters that won't be easily mistaken for each
+		// other. So, for example, omit o O and 0, 1 l and L.
+		String letters = "abcdefghjkmnpqrstuvwxyzABCDEFGHJKMNPQRSTUVWXYZ23456789+@";
+
+		String pw = "";
+		for (int i = 0; i < PASSWORD_LENGTH; i++) {
+			int index = (int) (RANDOM.nextDouble() * letters.length());
+			pw += letters.substring(index, index + 1);
+		}
+		return pw;
+	}
 
 	@RequestMapping(value = "/", method = RequestMethod.GET)
 	public ModelAndView test(HttpServletRequest request,
 			HttpServletResponse response) throws IOException {
-		if (request.isUserInRole("ROLE_ADMINISTRATOR")) {
-			ModelAndView view = new ModelAndView("securite/admin");
+
+		ModelAndView view = new ModelAndView();
+
+		if (request.isUserInRole("ROLE_ADMINISTRATOR")
+				|| request.isUserInRole("ROLE_HELPDESK")
+				|| request.isUserInRole("ROLE_MANAGER")) {
+			view.setViewName("securite/home");
 
 			List<User> users = new ArrayList<User>();
 			if (request.getParameter("fullname") == null
@@ -59,27 +87,31 @@ public class HomeController {
 			view.addObject("defaultView", defaultView);
 			view.addObject("errorMsg", errorMsg);
 
-			return view;
+			errorMsg = "";
+		} else {
+			view.setViewName("login");
 		}
-		return new ModelAndView("login");
-
+		return view;
 	}
 
 	@RequestMapping(value = "/addUser", method = RequestMethod.POST)
 	public String testaddUser(HttpServletRequest request,
 			HttpServletResponse response) throws IOException,
 			InvalidNameException {
+		if (request.isUserInRole("ROLE_ADMINISTRATOR")
+				|| request.isUserInRole("ROLE_HELPDESK")) {
+			errorMsg = "";
+			defaultView = "users";
 
-		errorMsg = "";
-		defaultView = "users";
+			User u = new User(request.getParameter("fullname"),
+					request.getParameter("surname"),
+					request.getParameter("password"));
+			try {
+				userService.create(u);
+			} catch (NamingException e) {
+				errorMsg = this.getErrorLDAP(e);
+			}
 
-		User u = new User(request.getParameter("fullname"),
-				request.getParameter("surname"),
-				request.getParameter("password"));
-		try {
-			userService.create(u);
-		} catch (NamingException e) {
-			errorMsg = this.getErrorLDAP(e);
 		}
 
 		return "redirect:/";
@@ -90,19 +122,23 @@ public class HomeController {
 			HttpServletResponse response) throws IOException,
 			InvalidNameException {
 
-		errorMsg = "";
-		defaultView = "groups";
+		if (request.isUserInRole("ROLE_ADMINISTRATOR")
+				|| request.isUserInRole("ROLE_MANAGER")) {
 
-		try {
-			User u = userService.findByCn(request.getParameter("fullname"));
+			errorMsg = "";
+			defaultView = "groups";
 
-			Group g = new Group(request.getParameter("groupname"));
-			g.setMembers(new HashSet<Name>());
-			g.addMember(u.getDn());
+			try {
+				User u = userService.findByCn(request.getParameter("fullname"));
 
-			groupService.create(g);
-		} catch (NamingException e) {
-			errorMsg = this.getErrorLDAP(e);
+				Group g = new Group(request.getParameter("groupname"));
+				g.setMembers(new HashSet<Name>());
+				g.addMember(u.getDn());
+
+				groupService.create(g);
+			} catch (NamingException e) {
+				errorMsg = this.getErrorLDAP(e);
+			}
 		}
 
 		return "redirect:/";
@@ -112,22 +148,25 @@ public class HomeController {
 	public String addUserToGroup(HttpServletRequest request,
 			HttpServletResponse response) throws IOException,
 			InvalidNameException {
+		if (request.isUserInRole("ROLE_ADMINISTRATOR")
+				|| request.isUserInRole("ROLE_HELPDESK")
+				|| request.isUserInRole("ROLE_MANAGER")) {
+			errorMsg = "";
+			defaultView = "groups";
 
-		errorMsg = "";
-		defaultView = "groups";
+			try {
+				User u = userService.findByCn(request.getParameter("fullname"));
 
-		try {
-			User u = userService.findByCn(request.getParameter("fullname"));
+				Group g = groupService.findByCn(request
+						.getParameter("groupname"));
 
-			Group g = groupService.findByCn(request.getParameter("groupname"));
+				g.addMember(u.getDn());
 
-			g.addMember(u.getDn());
-
-			groupService.update(g);
-		} catch (NamingException e) {
-			errorMsg = this.getErrorLDAP(e);
+				groupService.update(g);
+			} catch (NamingException e) {
+				errorMsg = this.getErrorLDAP(e);
+			}
 		}
-
 		return "redirect:/";
 	}
 
@@ -135,22 +174,27 @@ public class HomeController {
 	public String removeUserFromGroup(HttpServletRequest request,
 			HttpServletResponse response) throws IOException,
 			InvalidNameException {
+		if (request.isUserInRole("ROLE_ADMINISTRATOR")
+				|| request.isUserInRole("ROLE_HELPDESK")
+				|| request.isUserInRole("ROLE_MANAGER")) {
+			errorMsg = "";
+			defaultView = "groups";
 
-		errorMsg = "";
-		defaultView = "groups";
+			try {
+				// TODO@gent remove user from group
+				User u = userService.findByCn(request.getParameter("fullname"));
 
-		try {
-			User u = userService.findByCn(request.getParameter("fullname"));
+				Group g = groupService.findByCn(request
+						.getParameter("groupname"));
 
-			Group g = groupService.findByCn(request.getParameter("groupname"));
+				g.removeMember(u.getDn());
 
-			g.removeMember(u.getDn());
+				groupService.update(g);
 
-			groupService.update(g);
-		} catch (NamingException e) {
-			errorMsg = this.getErrorLDAP(e);
+			} catch (NamingException e) {
+				errorMsg = this.getErrorLDAP(e);
+			}
 		}
-
 		return "redirect:/";
 	}
 
@@ -158,18 +202,26 @@ public class HomeController {
 	public String deleteUser(HttpServletRequest request,
 			HttpServletResponse response) throws IOException,
 			InvalidNameException {
+		if (request.isUserInRole("ROLE_ADMINISTRATOR")) {
+			errorMsg = "";
+			defaultView = "users";
 
-		errorMsg = "";
-		defaultView = "users";
+			try {
+				User u = userService.findByCn(request.getParameter("fullname"));
 
-		try {
-			User u = userService.findByCn(request.getParameter("fullname"));
+				// TODO@gent remove user from group
+				List<Group> groups = groupService.findAll();
 
-			userService.delete(u);
-		} catch (NamingException e) {
-			errorMsg = this.getErrorLDAP(e);
+				for (Group group : groups) {
+					group.removeMember(u.getDn());
+					groupService.update(group);
+				}
+
+				userService.delete(u);
+			} catch (NamingException e) {
+				errorMsg = this.getErrorLDAP(e);
+			}
 		}
-
 		return "redirect:/";
 
 	}
@@ -178,18 +230,20 @@ public class HomeController {
 	public String deleteGroup(HttpServletRequest request,
 			HttpServletResponse response) throws IOException,
 			InvalidNameException {
+		if (request.isUserInRole("ROLE_ADMINISTRATOR")
+				|| request.isUserInRole("ROLE_MANAGER")) {
+			errorMsg = "";
+			defaultView = "groups";
 
-		errorMsg = "";
-		defaultView = "groups";
+			try {
+				Group g = groupService.findByCn(request
+						.getParameter("groupname"));
 
-		try {
-			Group g = groupService.findByCn(request.getParameter("groupname"));
-
-			groupService.delete(g);
-		} catch (NamingException e) {
-			errorMsg = this.getErrorLDAP(e);
+				groupService.delete(g);
+			} catch (NamingException e) {
+				errorMsg = this.getErrorLDAP(e);
+			}
 		}
-
 		return "redirect:/";
 
 	}
@@ -197,25 +251,25 @@ public class HomeController {
 	@RequestMapping(value = "/updateUser", method = RequestMethod.POST)
 	public String updateUser(HttpServletRequest request,
 			HttpServletResponse response) throws IOException {
+		if (request.isUserInRole("ROLE_ADMINISTRATOR")
+				|| request.isUserInRole("ROLE_HELPDESK")) {
+			errorMsg = "";
+			defaultView = "users";
 
-		errorMsg = "";
-		defaultView = "users";
+			try {
+				User u = userService.findByCn(request.getParameter("cn"));
 
-		try {
-			User u = userService.findByCn(request.getParameter("cn"));
+				u.setFullName(request.getParameter("fullname"));
+				u.setSurName(request.getParameter("surname"));
 
-			// TODO dn string to name
-			// u.setDn(request.getParameter("dn"));
-			u.setFullName(request.getParameter("fullname"));
-			u.setSurName(request.getParameter("surname"));
+				if (request.getParameter("password") != "") {
+					u.setPassword(request.getParameter("password"));
+				}
 
-			if (request.getParameter("password") != "") {
-				u.setPassword(request.getParameter("password"));
+				userService.update(u);
+			} catch (NamingException e) {
+				errorMsg = this.getErrorLDAP(e);
 			}
-
-			userService.update(u);
-		} catch (NamingException e) {
-			errorMsg = this.getErrorLDAP(e);
 		}
 
 		return "redirect:/";
@@ -225,22 +279,22 @@ public class HomeController {
 	@RequestMapping(value = "/updateGroup", method = RequestMethod.POST)
 	public String updateGroup(HttpServletRequest request,
 			HttpServletResponse response) throws IOException {
+		if (request.isUserInRole("ROLE_ADMINISTRATOR")
+				|| request.isUserInRole("ROLE_MANAGER")) {
+			errorMsg = "";
+			defaultView = "groups";
 
-		errorMsg = "";
-		defaultView = "groups";
+			try {
+				Group g = groupService.findByCn(request.getParameter("cn"));
 
-		try {
-			Group g = groupService.findByCn(request.getParameter("cn"));
+				// TODO@gent
+				g.setName(request.getParameter("groupname"));
 
-			// TODO dn string to name
-			// g.setDn(request.getParameter("dn"));
-			// g.setName(request.getParameter("groupname"));
-
-			groupService.update(g);
-		} catch (NamingException e) {
-			errorMsg = this.getErrorLDAP(e);
+				groupService.update(g);
+			} catch (NamingException e) {
+				errorMsg = this.getErrorLDAP(e);
+			}
 		}
-
 		return "redirect:/";
 
 	}
@@ -248,19 +302,21 @@ public class HomeController {
 	@RequestMapping(value = "/resetPassword", method = RequestMethod.GET)
 	public String resetPassword(HttpServletRequest request,
 			HttpServletResponse response) throws IOException {
+		if (request.isUserInRole("ROLE_ADMINISTRATOR")
+				|| request.isUserInRole("ROLE_HELPDESK")) {
+			errorMsg = "";
+			defaultView = "users";
 
-		errorMsg = "";
-		defaultView = "users";
-
-		try {
-			User u = userService.findByCn(request.getParameter("fullname"));
-
-			u.setPassword(defaultPassword);
-			userService.update(u);
-		} catch (NamingException e) {
-			errorMsg = this.getErrorLDAP(e);
+			try {
+				User u = userService.findByCn(request.getParameter("fullname"));
+				String password = generateRandomPassword();
+				u.setPassword(password);
+				userService.update(u);
+				errorMsg = "New password is : " + password;
+			} catch (NamingException e) {
+				errorMsg = this.getErrorLDAP(e);
+			}
 		}
-
 		return "redirect:/";
 
 	}
